Replace p4Repl() use of strtol() with a custom string to number parser p4StrNum().
strtol() assumes NUL terminated string, but the input source might be
read-only so not possible to terminate without copying to a buffer,
yet when the input source is a RAM the current numeric input might
overlap with any previous numeric input without the NUL terminator.
Need to convert based on the length of the parsed string.

Also the previous version did not allow for a possible negative sign
after $ # or % prefixes.